added validators

+++ ads/models.py
@@ -18,7 +18,7 @@ class Categories(models.Model):
 class Ads(models.Model):
     name = models.CharField(max_length=150, validators=[MinLengthValidator(10)])
     author = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.PositiveIntegerField()
     description = models.TextField(null=True)
     is_published = models.BooleanField(default=False)
     image = models.ImageField(upload_to='images/', null=True, blank=True)
